Reorganised files, modified functions and read more about header files and linking.
