+++ search_func/CMakeLists.txt
@@ -1,4 +1,7 @@
-file(GLOB SOURCE CONFIGURE_DEPENDS ${CMAKE_CURRENT_SOURCE_DIR}/Source/*.cpp)
-add_executable(searchTest ${SOURCE})
+set(SOURCE ${CMAKE_CURRENT_SOURCE_DIR}/Source/search_func.cpp 
+    ${CMAKE_CURRENT_SOURCE_DIR}/Source/main.cpp 
+    ${CMAKE_CURRENT_SOURCE_DIR}/Source/utils.cpp)
+    
+add_executable(searchTest ${SOURCE} ${CMAKE_CURRENT_SOURCE_DIR}/../DeckBuilder/Card.cpp)
 include_directories(${CMAKE_CURRENT_SOURCE_DIR}/Include ${CMAKE_CURRENT_SOURCE_DIR}/../DeckBuilder)
 target_link_libraries(searchTest ${SQLite3_LIBRARIES})
