+++ search_func/Include/utils.hpp
@@ -1,0 +1,7 @@
+#pragma once
+#include <vector>
+#include <string>
+
+
+unsigned int find(std::vector<std::string>, std::string);
+std::string toString(std::size_t);
